Have scene insertion return the id

+++ dataLayer/src/test/java/com/sildian/apps/storywriter/datalayer/scene/RandomSceneDb.kt
@@ -4,7 +4,7 @@ import com.sildian.apps.storywriter.domainlayer.nextString
 import kotlin.random.Random
 
 internal fun Random.nextSceneDb(
-    id: Int = nextInt(from = 1, until = 100),
+    id: Long = nextLong(from = 1, until = 100),
     description: String = nextString(),
 ): SceneDb = SceneDb(
     id = id,
